requirements.txt
asgiref==3.7.2
Django==5.0.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.4
sqlparse==0.4.4
whitenoise==6.6.0
